# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for Survey Application",
        "title": "Survey",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/add-ans": {
            "post": {
                "description": "Adds a new answer to the survey",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "answer"
                ],
                "parameters": [
                    {
                        "description": "addAnswer input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/addAnswer.Input"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.Answer"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/delete-ans/{survey_id}/{ans_id}": {
            "delete": {
                "description": "Deletes an answer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "answer"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Survey ID",
                        "name": "survey_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Answer ID",
                        "name": "ans_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/delete-survey/{survey_id}": {
            "delete": {
                "description": "Deletes a survey",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "survey"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Survey ID",
                        "name": "survey_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/get-result": {
            "post": {
                "description": "Gets a result of the survey",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "survey"
                ],
                "parameters": [
                    {
                        "description": "Survey ID",
                        "name": "survey_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/getResult.Input"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.Survey"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/start-survey": {
            "post": {
                "description": "Starts a new survey",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "survey"
                ],
                "parameters": [
                    {
                        "description": "startSurvey input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/startSurvey.Input"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.Survey"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/vote": {
            "post": {
                "description": "Votes for the answer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "answer"
                ],
                "parameters": [
                    {
                        "description": "Vote input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vote.Input"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "addAnswer.Input": {
            "type": "object",
            "properties": {
                "survey_id": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "entities.Answer": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "surveyID": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                },
                "votes": {
                    "type": "integer"
                }
            }
        },
        "entities.Response": {
            "type": "object",
            "properties": {
                "message": {}
            }
        },
        "entities.Survey": {
            "type": "object",
            "properties": {
                "answers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Answer"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "getResult.Input": {
            "type": "object",
            "properties": {
                "survey_id": {
                    "type": "integer"
                }
            }
        },
        "startSurvey.Input": {
            "type": "object",
            "properties": {
                "title": {
                    "type": "string"
                }
            }
        },
        "vote.Input": {
            "type": "object",
            "properties": {
                "answer_id": {
                    "type": "integer"
                },
                "survey_id": {
                    "type": "integer"
                }
            }
        }
    }
}